+++ lib/auth/reset_password.dart
@@ -99,8 +99,16 @@ class _ResetPasswordForm extends State<ResetPasswordWidget> {
                 width: 300.0,
                 height: 400.0,
                 decoration: BoxDecoration(
-                  color: const Color.fromARGB(255, 235, 235, 235),
+                  color: const Color.fromARGB(255, 245, 245, 245),
                   borderRadius: BorderRadius.circular(10),
+                  boxShadow: [
+                    BoxShadow(
+                      color: Color.fromARGB(50,0,0,0),
+                      blurRadius: 15,
+                      spreadRadius: 5,
+                      offset: Offset(0, 15)
+                    )
+                  ]
                 ),
                 child: Column(
                   children: [
@@ -129,7 +137,7 @@ class _ResetPasswordForm extends State<ResetPasswordWidget> {
                             borderSide: BorderSide.none,
                           ),
                           filled: true,
-                          fillColor: const Color.fromARGB(10, 0, 0, 0)
+                          fillColor: const Color.fromARGB(5, 0, 0, 0)
                         ),
                         validator: (value) {
                           if(value!.isEmpty) {
@@ -152,7 +160,7 @@ class _ResetPasswordForm extends State<ResetPasswordWidget> {
                             borderSide: BorderSide.none,
                           ),
                           filled: true,
-                          fillColor: const Color.fromARGB(10, 0, 0, 0)
+                          fillColor: const Color.fromARGB(5, 0, 0, 0)
                         ),
                         validator: (value) {
                           if(value!.isEmpty) {
@@ -171,7 +179,7 @@ class _ResetPasswordForm extends State<ResetPasswordWidget> {
                           backgroundColor: isPressed ? const Color.fromARGB(255, 83, 83, 83) : Colors.blue,
                           foregroundColor: Colors.white,
                         ),
-                        child: Text("Mandar reseteo"),
+                        child: Text("Recuperar contraseña"),
                       ),
                     ),
                     SizedBox(height: 20.0,),